Simple prototype that merges textures that look like occupancy grid maps.

+++ glamm/frame_buffer.cpp
@@ -32,7 +32,7 @@ FrameBuffer::FrameBuffer(const unsigned int width, const unsigned int height)
   glBindTexture(GL_TEXTURE_2D, this->texture_id_);
 
   glTexImage2D(
-    GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+    GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
   glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
   glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
   glBindTexture(GL_TEXTURE_2D, 0);
